+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixes Kafka::FFI::Client#cluster_id being incorrectly bound
 * Fixes naming issues in Message#headers and Message#detach_headers
 * Fixes naming issue in Config#set_ssl_cert
+* Fixes passing nil for a Kafka::FFI::Opaque
 * Adds all current RD_KAFKA_RESP_ERR_ constants
 * Adds Kafka::QueueFullError for when producer queue is full